refactor(folder): 서비스와 훅 레이어 책임 정리
